refactor(opus): move per-arch logic out of gen_instances.py
Codegen entry point shrinks from 1858 -> 1221 lines. All arch-specific
state (maps, MFMA shape sets, a16w16 validators, LDS budget) now lives in
its arch module; gen_instances.py keeps only cross-arch helpers, the
emit-dispatch router, and per-pass TU emission.

- codegen/common.py: cross-arch tag groupings + WARP_SIZE + ARCH_MAP_REGISTRY
  helpers; no arch-specific code left.
- codegen/gen_instances_gfx950.py: PIPELINE/TRAITS/KERNEL_FUNC/TRAITS_NAME/
  KARGS maps (+ 4g_safe), 5 VALID_*_MFMA, _validate_a16w16 (split-barrier),
  _validate_a16w16_flatmm{,_splitk}, _validate_a16w16_persistent,
  _validate_a16w16_mono_tile. Each map self-registers via register_arch_map.
- codegen/gen_instances_gfx942.py: gfx942 maps + VALID_GFX942_BF16_MFMA +
  _validate_a16w16_gfx942 (minimal physical-limit check: MFMA shape,
  BLOCK<=512, AGPR/VGPR caps, LDS lower-bound vs 64 KiB CDNA3 budget).
  Detailed depth/padding checks remain in gfx942/*.cuh static_asserts.
- gen_instances.py: removes 5 maps + 5 VALID_*_MFMA + 5 validators + the
  4 redundant GFX942_*_MAP override dicts (3 were no-ops, 1 was working
  around a missing default entry now folded into the gfx942 map). Validator
  dispatch routes by tag set (a16w16 -> gfx950; _GFX942_A16W16_TAGS ->
  gfx942) without an arch_prefix branch. Also surfaces 2 previously-silent
  splitk tags (kbuf2v_sk / kbuf2v_bk128_sk) in the validator log.

Generated .cu/.cuh are byte-identical (diff -rq = 0). Validated:
- multi-arch build (gfx942;gfx950 PREBUILD_KERNELS=1): clean link.
- op_tests/opus/run_tests.sh device suite: all PASS.
- a16w16 parity sweep (50 shapes, all kid families): 50/50 PASS, err=0.
- black + ruff: clean.
