+++ csrc/opus_gemm/codegen/common.py
@@ -9,7 +9,46 @@ dispatch_emit(cg, k, **kwargs).  Adding a new arch (e.g. gfx1250) = one new
 file + one new import; entry point itself is arch-agnostic.
 """
 
+WARP_SIZE = 64
+
+# Paired W3 kernels (nosplit_tag -> splitk_tag) share one <Traits, Kargs> template.
+W3_KERNEL_PAIRS = {
+    "a16w16_kbuf3": "a16w16_kbuf3_sk",
+    "a16w16_kbuf2v": "a16w16_kbuf2v_sk",
+    "a16w16_kbuf2v_bk128": "a16w16_kbuf2v_bk128_sk",
+    "a16w16_kbuf1": "a16w16_kbuf1_sk",
+}
+_NOSPLIT = tuple(W3_KERNEL_PAIRS.keys())
+_SPLITK = tuple(W3_KERNEL_PAIRS.values())
+_GFX942_A16W16_TAGS = (
+    _SPLITK + ("a16w16_fused_reduce", "a16w16_kbuf1_large_tile") + _NOSPLIT
+)
+_A16W16_TAGS = (
+    "a16w16",
+    "a16w16_flatmm",
+    "a16w16_flatmm_splitk",
+    "a16w16_persistent",
+    "a16w16_mono_tile",
+) + _GFX942_A16W16_TAGS
+
 EMIT_REGISTRY = {}
+
+# Per-arch map registry: {(arch, map_name): dict}. Each arch module registers
+# its overrides at import time; gen_instances merges them into the cross-arch
+# default maps.
+ARCH_MAP_REGISTRY = {}
+
+
+def register_arch_map(arch, map_name, mapping):
+    key = (arch, map_name)
+    if key in ARCH_MAP_REGISTRY:
+        raise RuntimeError(f"arch map already registered for {key}")
+    ARCH_MAP_REGISTRY[key] = mapping
+
+
+def get_arch_map(arch, map_name):
+    """Return the registered map, or {} if none."""
+    return ARCH_MAP_REGISTRY.get((arch, map_name), {})
 
 
 def kid_arch(k):
